Feat : UserDataProvider 분리 - API 연결 테스트중 user관리 객체의 필요로 구현 - UserDataProviding으로 추상화
Issue #119
